Rename index1 to index, delete unnecessary html file
